requirements.txt
python-dotenv
openai
pandas
# rpy2
pyreadr
databases[sqlite]